split quadtree into individual files for better organiztion
