+++ src/quadtree/mod.rs
@@ -1,0 +1,10 @@
+pub mod node;
+pub mod quad;
+pub mod traits;
+
+pub use node::Node;
+//pub use quad::Quad;
+//pub use traits::*;
+
+mod quadtree;
+pub use quadtree::Quadtree;
