+++ src/quadtree/traits.rs
@@ -1,0 +1,4 @@
+/*pub trait Spatial {
+    fn position(&self) -> ultraviolet::Vec2;
+    // Add more trait methods as needed
+}*/
